fix: fallback error message
