+++ packages/event-target/typed-event-target/src/index.ts
@@ -1,5 +1,3 @@
-// extends EventTarget types and adds typed Events
-
 export type TypedEvent = Event & {
   data?: unknown
 }
@@ -30,4 +28,40 @@ export interface TypedEventTarget<TEvent extends TypedEvent>
     callback: TypedEventListenerOrEventListenerObject<TEvent>,
     options?: EventListenerOptions | boolean,
   ): void
+}
+
+export type EventAPI = {
+  Event: typeof Event
+  EventTarget: typeof EventTarget
+}
+
+export const resolveGlobalConfig = (api: EventAPI = globalThis): EventAPI => {
+  if (!api.Event || !api.EventTarget) {
+    throw new Error(`
+[@graphql-yoga/subscription] 'createPubSub' uses the Event and EventTarget APIs.
+
+In modern JavaScript environments those are part of the global scope. However, if you are using an older version of Node.js (< 16.x.x), those APIs must be polyfilled.
+You can provide polyfills to the 'createPubSub' function:
+
+\`\`\`
+// yarn install --exact event-target-polyfill@0.0.3
+import 'event-target-polyfill'
+
+const pubSub = createPubSub()
+\`\`\`
+
+Alternatively, you can provide your own custom implementation.
+
+\`\`\`
+const pubSub = createPubSub({
+  event: {
+    Event,
+    EventTarget,
+  }
+})
+\`\`\`
+`)
+  }
+
+  return globalThis
 }
